Adding correct padding to info list
