+++ lib/app/shared/widgets/layout/custom_scaffold/custom_scaffold.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:my_services/app/shared/themes/themes.dart';
+import 'package:my_services/app/shared/widgets/layout/custom_scaffold/widgets/padding_wrap.dart';
 import 'package:my_services/app/shared/widgets/layout/custom_scroll_behavior/custom_scroll_behavior.dart';
 
 class CustomScaffold extends StatelessWidget {
@@ -7,10 +7,12 @@ class CustomScaffold extends StatelessWidget {
     super.key,
     this.onRefresh,
     this.child,
+    this.padding,
   });
 
   final Future<void> Function()? onRefresh;
   final Widget? child;
+  final EdgeInsets? padding;
 
   @override
   Widget build(BuildContext context) {
@@ -20,27 +22,16 @@ class CustomScaffold extends StatelessWidget {
         child: onRefresh != null
             ? RefreshIndicator(
                 onRefresh: onRefresh!,
-                child: DefaultPadding(child: child),
+                child: PaddingWrap(
+                  padding: padding,
+                  child: child,
+                ),
               )
-            : DefaultPadding(child: child),
+            : PaddingWrap(
+                padding: padding,
+                child: child,
+              ),
       ),
-    );
-  }
-}
-
-class DefaultPadding extends StatelessWidget {
-  const DefaultPadding({super.key, this.child});
-  final Widget? child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(
-        left: AppSizeConstants.largeSpace,
-        right: AppSizeConstants.largeSpace,
-        top: AppSizeConstants.largeSpace,
-      ),
-      child: child,
     );
   }
 }
