+++ lib/app/shared/widgets/layout/custom_scaffold/widgets/padding_wrap.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+import 'package:my_services/app/shared/themes/themes.dart';
+
+class PaddingWrap extends StatelessWidget {
+  const PaddingWrap({
+    super.key,
+    this.child,
+    this.padding,
+  });
+
+  final Widget? child;
+  final EdgeInsets? padding;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: padding ??
+          const EdgeInsets.only(
+            left: AppSizeConstants.largeSpace,
+            right: AppSizeConstants.largeSpace,
+            top: AppSizeConstants.largeSpace,
+          ),
+      child: child,
+    );
+  }
+}
