+++ lib/app/views/services/service_landing/widgets/service_navbar.dart
@@ -23,58 +23,54 @@ class ServiceNavbar extends StatelessWidget {
   Widget build(BuildContext context) {
     final serviceCubit = context.read<ServiceLandingCubit>();
 
-    return Padding(
-      padding:
-          const EdgeInsets.symmetric(horizontal: AppSizeConstants.smallSpace),
-      child: TextWithTrailing(
-        text: context.appLocalizations.services,
-        trailing: Row(
-          children: [
-            CircularButton(
-              onTap: () => showModalBottomSheet(
-                context: context,
-                useRootNavigator: true,
-                isScrollControlled: true,
-                builder: (context) => OrderByBottomSheet(
-                  selectedOption: serviceCubit.state.selectedOrderBy,
-                  onPressed: (orderBy) {
-                    context.pop();
-                    serviceCubit.onChangeOrderBy(orderBy);
-                  },
-                ),
-              ),
-              child: const Icon(
-                Icons.swap_vert,
-                size: 18,
+    return TextWithTrailing(
+      text: context.appLocalizations.services,
+      trailing: Row(
+        children: [
+          CircularButton(
+            onTap: () => showModalBottomSheet(
+              context: context,
+              useRootNavigator: true,
+              isScrollControlled: true,
+              builder: (context) => OrderByBottomSheet(
+                selectedOption: serviceCubit.state.selectedOrderBy,
+                onPressed: (orderBy) {
+                  context.pop();
+                  serviceCubit.onChangeOrderBy(orderBy);
+                },
               ),
             ),
-            AppSizeConstants.smallHorizontalSpacer,
-            CircularButton(
-              showCircularIndicator: serviceCubit.state.didFiltersChange,
-              onTap: () => showModalBottomSheet(
-                context: context,
-                useRootNavigator: true,
-                isScrollControlled: true,
-                builder: (context) => FiltersBottomSheet(
-                  dateKey: dateKey,
-                  dateController: dateController,
-                ),
-              ),
-              child: const Icon(
-                Icons.filter_list_alt,
-                size: 18,
+            child: const Icon(
+              Icons.swap_vert,
+              size: 18,
+            ),
+          ),
+          AppSizeConstants.smallHorizontalSpacer,
+          CircularButton(
+            showCircularIndicator: serviceCubit.state.didFiltersChange,
+            onTap: () => showModalBottomSheet(
+              context: context,
+              useRootNavigator: true,
+              isScrollControlled: true,
+              builder: (context) => FiltersBottomSheet(
+                dateKey: dateKey,
+                dateController: dateController,
               ),
             ),
-            AppSizeConstants.smallHorizontalSpacer,
-            PillButton(
-              onTap: () {
-                context.read<AppCubit>().changeToAddServicePage();
-                context.go(AppRoutes.addServices);
-              },
-              child: Text(context.appLocalizations.newService),
-            )
-          ],
-        ),
+            child: const Icon(
+              Icons.filter_list_alt,
+              size: 18,
+            ),
+          ),
+          AppSizeConstants.smallHorizontalSpacer,
+          PillButton(
+            onTap: () {
+              context.read<AppCubit>().changeToAddServicePage();
+              context.go(AppRoutes.addServices);
+            },
+            child: Text(context.appLocalizations.newService),
+          )
+        ],
       ),
     );
   }
